MISC - Fixed polkadot balance
