+++ frontend/src/locales/lang/en/routes/settings.ts
@@ -1,8 +1,3 @@
 export default {
   settings: 'Settings',
-  pleaseInputWsUrl: 'Please input WsUrl',
-  address: 'Address',
-  accountBalance: 'Account Balance',
-  save: 'Save',
-  unbind: 'Unbind',
 };
