+++ frontend/src/locales/lang/en/settings.ts
@@ -1,0 +1,10 @@
+export default {
+  index: {
+    settings: 'Settings',
+    pleaseInputWsUrl: 'Please input WsUrl',
+    address: 'Address',
+    accountBalance: 'Account Balance',
+    save: 'Save',
+    unbind: 'Unbind',
+  },
+};
